refactor: 토큰 관련 로직 User와 분리하여 Auth로 개편
