+++ src/main/java/team/silvertown/masil/auth/repository/RefreshTokenRepository.java
@@ -1,7 +1,7 @@
-package team.silvertown.masil.token.repository;
+package team.silvertown.masil.auth.repository;
 
 import org.springframework.data.repository.CrudRepository;
-import team.silvertown.masil.token.domain.RefreshToken;
+import team.silvertown.masil.auth.domain.RefreshToken;
 
 public interface RefreshTokenRepository extends CrudRepository<RefreshToken, String> {
