+++ src/main/java/team/silvertown/masil/config/jwt/JwtTokenProvider.java
@@ -22,7 +22,7 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
 import team.silvertown.masil.user.domain.Authority;
-import team.silvertown.masil.user.dto.LoginResponse;
+import team.silvertown.masil.auth.dto.LoginResponse;
 
 @Slf4j
 @Component
