+++ src/main/java/team/silvertown/masil/user/controller/UserController.java
@@ -5,45 +5,32 @@ import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
-import io.swagger.v3.oas.annotations.security.SecurityRequirements;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.servlet.http.HttpServletRequest;
-import java.util.Objects;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
-import team.silvertown.masil.security.exception.InvalidAuthenticationException;
-import team.silvertown.masil.user.dto.LoginResponse;
 import team.silvertown.masil.user.dto.MeInfoResponse;
 import team.silvertown.masil.user.dto.MyPageInfoResponse;
 import team.silvertown.masil.user.dto.NicknameCheckResponse;
 import team.silvertown.masil.user.dto.OnboardRequest;
 import team.silvertown.masil.user.dto.UpdateRequest;
 import team.silvertown.masil.user.dto.UpdateResponse;
-import team.silvertown.masil.user.exception.UserErrorCode;
 import team.silvertown.masil.user.service.UserService;
 
 @RestController
 @RequiredArgsConstructor
 @Tag(name = "회원 관련 API")
 public class UserController {
-
-    private static final String ACCESS_TOKEN_PREFIX = "Bearer ";
-    private static final String REFRESH_TOKEN_HEADER = "Refresh-Token";
 
     private final UserService userService;
 
@@ -92,55 +79,6 @@ public class UserController {
         Long memberId
     ) {
         return ResponseEntity.ok(userService.getMe(memberId));
-    }
-
-    @PostMapping("/api/v1/users/login")
-    @Operation(summary = "카카오 토큰으로 로그인")
-    @ApiResponse(
-        responseCode = "200",
-        content = @Content(
-            mediaType = "application/json",
-            schema = @Schema(implementation = LoginResponse.class)
-        )
-    )
-    @SecurityRequirements(
-        @SecurityRequirement(name = "토큰 받아오기")
-    )
-    public ResponseEntity<LoginResponse> login(
-        @RequestHeader(HttpHeaders.AUTHORIZATION)
-        @Parameter(hidden = true)
-        String accessToken
-    ) {
-        return ResponseEntity.ok(userService.login(accessToken));
-    }
-
-    @GetMapping("api/v1/users/login/refresh")
-    @Operation(summary = "리프레시 토큰으로 새 토큰 받기")
-    @SecurityRequirements(
-        {
-            @SecurityRequirement(name = "카카오 토큰 로그인"),
-            @SecurityRequirement(name = "리프레시 토큰")
-        }
-    )
-    public ResponseEntity<Void> refresh(
-        HttpServletRequest request
-    ) {
-        String refreshToken = validateExistHeader(request);
-        String newToken = userService.refresh(refreshToken);
-
-        return ResponseEntity.ok()
-            .header(HttpHeaders.AUTHORIZATION, ACCESS_TOKEN_PREFIX + newToken)
-            .build();
-    }
-
-    private String validateExistHeader(HttpServletRequest request) {
-        String authorizationHeader = request.getHeader(HttpHeaders.AUTHORIZATION);
-        String refreshTokenHeader = request.getHeader(REFRESH_TOKEN_HEADER);
-        if (Objects.isNull(authorizationHeader) || Objects.isNull(refreshTokenHeader)) {
-            throw new InvalidAuthenticationException(UserErrorCode.INVALID_JWT_TOKEN);
-        }
-
-        return request.getHeader("Refresh-Token");
     }
 
     @PutMapping(
