+++ src/main/java/team/silvertown/masil/user/service/KakaoOAuthService.java
@@ -7,8 +7,10 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import team.silvertown.masil.security.exception.InvalidAuthenticationException;
 import team.silvertown.masil.user.dto.KakaoResponse;
 import team.silvertown.masil.user.dto.OAuthResponse;
+import team.silvertown.masil.user.exception.UserErrorCode;
 import team.silvertown.masil.user.exception.UserValidator;
 import team.silvertown.masil.user.service.restTemplate.RestTemplateService;
 
@@ -22,21 +24,27 @@ public class KakaoOAuthService implements OAuthService {
     private final ObjectMapper objectMapper;
     private final RestTemplateService restTemplateService;
 
-    public OAuthResponse getUserInfo(String kakaoToken) throws IOException {
+    public OAuthResponse getUserInfo(String kakaoToken){
         HashMap<String, Object> userInfo = new HashMap<>();
 
-        ResponseEntity<String> accountInfoResponse = restTemplateService.requestKaKaoInfo(
-            kakaoToken);
+        try {
+            ResponseEntity<String> accountInfoResponse = restTemplateService.requestKaKaoInfo(
+                kakaoToken);
 
-        String json = accountInfoResponse.getBody();
-        KakaoResponse kakaoResponse = objectMapper.readValue(json, KakaoResponse.class);
-        String id = kakaoResponse.getId();
-        UserValidator.validateProvidedId(id);
+            String json = accountInfoResponse.getBody();
+            KakaoResponse kakaoResponse = objectMapper.readValue(json, KakaoResponse.class);
+            String id = kakaoResponse.getId();
+            UserValidator.validateProvidedId(id);
 
-        userInfo.put("id", id);
-        userInfo.put("provider", KAKAO_PROVIDER);
+            userInfo.put("id", id);
+            userInfo.put("provider", KAKAO_PROVIDER);
 
-        return OAuthResponse.from(userInfo);
+            return OAuthResponse.from(userInfo);
+        } catch (Exception e){
+            log.error("social login error occured, the reason is: {}", e.getMessage(), e);
+            throw new InvalidAuthenticationException(UserErrorCode.INVALID_OAUTH2_TOKEN);
+        }
+
     }
 
 }
