+++ src/test/java/team/silvertown/masil/user/controller/UserControllerTest.java
@@ -1,0 +1,11 @@
+package team.silvertown.masil.user.controller;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+
+class UserControllerTest {
+
+}
